requirements.txt
djangorestframework
bs4
cssmin
django==3.0.4
django-assets
django-braces
django-drf-filepond==0.2.2
django-ipware
django-rest-framework==0.1.0
django-tinymce
geopy
gunicorn
jinja2
jsmin
pillow
psycopg2-binary
pytz
pyyaml
requests
sentry-sdk==0.14.3
webassets
